WebIDL/tests/submissions/heycam/operation-001.html

<!DOCTYPE html>
<title>Regular operations</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<script>
// 4.5.7 Operations
//
// Regular operations correspond to properties on the interface prototype
// object, which are writable, enumerable and configurable.  If they're on
// the [Global] object, then the properties will be on the instance instead.
// The .length property on the Function object is the shortest valid argument
// for the operation.

var operations = [

  // feature: HTMLElement.click (operation)
  "HTMLElement", "click", 0,

  // feature: Node.appendChild (operation)
  "Node", "appendChild", 1,

  // feature: DOMTokenList.add (operation) (variadic)
  "DOMTokenList", "add", 0,

  // feature: Window.setTimeout (operation) (variadic)
  "Window", "setTimeout", 1,

  // feature: Window.postMessage (operation) (with optional argument)
  "Window", "postMessage", 2,

  // feature: DOMImplementation.createHTMLDocument (operation) (with optional argument)
  "DOMImplementation", "createHTMLDocument", 0,

  // feature: Node.cloneNode (operation) (with optional argument default value)
  "Node", "cloneNode", 0,

  // feature: HTMLSelectElement.add (operation) (with optional argument default value)
  "HTMLSelectElement", "add", 1,

  // feature: CanvasRenderingContext2D.drawImage (operation) (overloaded)
  "CanvasRenderingContext2D", "drawImage", 3,

  // feature: FormData.append (operation) (overloaded)
  "FormData", "append", 2
];

function operationHome(intf) {
}

function operationHomeName(intf) {
}

for (var i = 0; i < operations.length; i+= 3) {
  var intf = operations[i];
  var name = operations[i + 1];
  var length = operations[i + 2];
  var object = interfaceMemberHome(intf);
  var objectName = interfaceMemberHomeName(intf);

  test(function() {
    assert_property(object, name, { writable: true, enumerable: true, configurable: true });
  }, "the " + objectName + " " + name + " property has the right attributes");

  test(function() {
    assert_equals(typeof window[intf].prototype[name], "function");
  }, objectName + " " + name + " property is a function");

  test(function() {
    assert_property(window[intf].prototype[name], "length", { writable: false, enumerable: false, configurable: false, value: length });
  }, objectName + " " + name + ".length  has correct length property");
}
</script>
